% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fasta.R
\name{check_error_tag_fasta}
\alias{check_error_tag_fasta}
\title{check_error_tag_fasta}
\usage{
check_error_tag_fasta(
  fasta_list = fasta_list,
  tag_site = tag_site,
  target_info = target_info,
  target = target,
  csv = csv
)
}
\arguments{
\item{fasta_list}{a list of lists, each list containing two elements
the name, first element, and a sequence referring to a genomic regions, second
element. This list can be created using \code{get_fasta}.}

\item{tag_site}{a integer or a numerical vector, representing the localization of the
tag site in the targets. If it is one number, the tags will be inserted at position
tag_site from the first nucleotide of target. if it is a numerical vector,
it should be the same length as the \code{fasta_list}. default = \code{round(nchar(target)/2)}.}

\item{target_info}{target info should be NULL if target is not given by a csv file.}

\item{target}{a character vector or dataframe containing the target region for each of the
sequences in the \code{fasta_list}. The length should be the same of \code{fasta_list}.
if target is a dataframe, it should have a column named "target".}

\item{csv}{a path to the csv file containing the list of target region or
table containing the target sequence. It will only be
considered if target is NULL.}
}
\value{
no value
}
\description{
check_error_tag_fasta
}
