% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord_website.R
\name{get_coord_website}
\alias{get_coord_website}
\title{GetCoordWebsite}
\usage{
get_coord_website(coord_table = coord_table, flank_n = 250)
}
\arguments{
\item{coord_table}{a dataframe or a .csv file of genomics coordinates
containing the columns:
"Name" - a reference name or tag for each of the coordinates,
'NCBI_ID' - the NCBI reference sequence ID (e.g. NC_010443.5),
"start" - the coordinate for the start position ,
"end" - the coordinate for the end position.}

\item{flank_n}{the number of nucleotides flanking upstream and downstream the
coordinates given (default = 250).}
}
\value{
listweb a list of URLs containing the fasta sequence of the coordinates.
}
\description{
This function uses a dataframe of genomic coordinates such as:
NC_010443.5 39157405 39157428 to get the NCBI URL of the fasta of the region plus
250nt upstream and downstream of the coordinates.
}
