% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fasta.R
\name{get_tag_seq}
\alias{get_tag_seq}
\title{get_tag_seq}
\usage{
get_tag_seq(fasta_list = fasta_list, loc_list = loc_list)
}
\arguments{
\item{fasta_list}{a list of lists, each list containing two elements
the name, first element, and a sequence referring to a genomic regions, second
element. This list can be created using \code{get_fasta}}

\item{loc_list}{list of locations calculated in the the tag_fasta}
}
\value{
a list of dinucleotide sequences
}
\description{
function used to extract the sequence of the position tag_site in the target on the
complete sequence
}
