% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fasta.R
\name{insert_tag}
\alias{insert_tag}
\title{insert_tag}
\usage{
insert_tag(
  fasta_list = fasta_list,
  loc_list = loc_list,
  tag_seq_list = tag_seq_list
)
}
\arguments{
\item{fasta_list}{a list of lists, each list containing two elements
the name, first element, and a sequence referring to a genomic regions, second
element. This list can be created using \code{get_fasta}}

\item{loc_list}{list of locations calculated in the the tag_fasta}

\item{tag_seq_list}{a list of dinucleotide sequences}
}
\value{
the fasta list containing the sequence with and without the tag
}
\description{
insert_tag
}
