% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fasta.R
\name{tag_fasta}
\alias{tag_fasta}
\title{tag_fasta}
\usage{
tag_fasta(fasta_list = fasta_list, target = NULL, csv = NULL, tag_site = NULL)
}
\arguments{
\item{fasta_list}{a list of lists, each list containing two elements
the name, first element, and a sequence referring to a genomic regions, second
element. This list can be created using \code{get_fasta}.}

\item{target}{a character vector or dataframe containing the target region for each of the
sequences in the \code{fasta_list}. The length should be the same of \code{fasta_list}.
if target is a dataframe, it should have a column named "target".}

\item{csv}{a path to the csv file containing the list of target region or
table containing the target sequence. It will only be
considered if target is NULL.}

\item{tag_site}{a integer or a numerical vector, representing the localization of the
tag site in the targets. If it is one number, the tags will be inserted at position
tag_site from the first nucleotide of target. if it is a numerical vector,
it should be the same length as the \code{fasta_list}. default = \code{round(nchar(target)/2)}.}
}
\value{
a list containing the fasta_list plus a sequence containing the tag for
each of the elements of the fasta_list.
}
\description{
this function marks (inserts []) a specific position
based on a target sequence on genomic sequences in a list. the [] are going to
be inserted at position \code{tag_site} and \code{tag_site}+ 1.
}
\examples{

tag_fasta(
fasta_list = list(
  list("target1", "aaatattttggt"),
  list("target2", "ggtgtaatgat")
),
target = data.frame(target = c("aaatatt", "gtaatg"), cut = c(3, 5)),
tag_site = 4
)

}
\seealso{
get_fasta
}
