% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{write_fasta}
\usage{
write_fasta(
  fasta_list = fasta_list,
  named = FALSE,
  contain_flag = TRUE,
  tagged = FALSE,
  combined = TRUE,
  path_to_file = path_to_file,
  width = 60,
  append = FALSE
)
}
\arguments{
\item{fasta_list}{a list of lists, each list containing at least two elements
the name, first element, and a sequence referring to a genomic regions, second
element. there could be a third element that would be the sequence including (tagging)
a target region in the sequence. The list can be named, and the name can be used to
create the header of the .fasta instead of the name contained in the first element.
the list can be created by the \code{get_fasta} function.}

\item{named}{logical if TRUE the header of the region in the .fasta file will be created
using the names of the list. When FALSE, the header will be creating from the first
element of each internal list. It overwrites the \code{contain_flag} to FALSE.}

\item{contain_flag}{if the name on the first elements of the lists already contain
the .fasta header tag ">" .}

\item{tagged}{logical if TRUE the sequence written in the .fasta file will be taken
from the third element of each internal list. If FALSE, the sequence will be taken
from second element. The \code{tag_fasta} file can create a tagged sequence.}

\item{combined}{logical if TRUE will create one file containing all sequences in the
list. If false will create one file for each sequence. The files names and location will
be given by the \code{path_to_file} that should be a character vector with the same length as
the \code{fasta_list}.}

\item{path_to_file}{a character vector of length one, if \code{combined} = TRUE, or length
equal to the length of \code{fasta_list} if \code{combined} = FALSE.}

\item{width}{the number of characters to be printed per line.
if not between 1 and 255, width with be infinite.}

\item{append}{\code{logical} indicating if the output should be
appended to \code{file} instead of overwriting it}
}
\value{
.fasta files.
}
\description{
A function to write .fasta files from a list of sequences.
}
\seealso{
get_fasta tag_fasta tigger::writeFasta
}
