% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta.R
\name{get_fasta}
\alias{get_fasta}
\title{get_fasta}
\usage{
get_fasta(
  weblist = NULL,
  coord_table = NULL,
  flank_n = 250,
  feature = "//pre",
  verbose = TRUE,
  port = netstat::free_port(),
  chromever = "105.0.5195.52",
  check = TRUE,
  browser = c("chrome", "firefox", "phantomjs", "internet explorer"),
  version = "latest",
  geckover = "latest"
)
}
\arguments{
\item{weblist}{a list of URLs containing the NCBI website of the fasta sequence
of genomic regions.}

\item{coord_table}{a dataframe or a .csv file of genomics coordinates
containing the columns:
"Name" - a reference name or tag for each of the coordinates,
'NCBI_ID' - the NCBI reference sequence ID (e.g. NC_010443.5),
"start" - the coordinate for the start position ,
"end" - the coordinate for the end position.}

\item{flank_n}{the number of nucleotides flanking upstream and downstream the
coordinates given (default = 250).}

\item{feature}{vector same length of the web_list or character string
of the html feature/s containing the fasta information in the NCBI html.}

\item{verbose}{If TRUE, include status messages (if any)}

\item{port}{Port to run on}

\item{chromever}{what version of Chrome driver to run. Default = "latest"
which runs the most recent version. To see other version currently
sourced run binman::list_versions("chromedriver"), A value of NULL
excludes adding the chrome browser to Selenium Server.}

\item{check}{If TRUE check the versions of selenium available and the
versions of associated drivers (chromever, geckover, phantomver,
iedrver). If new versions are available they will be downloaded.}

\item{browser}{Which browser to start}

\item{version}{what version of Selenium Server to run. Default = "latest"
which runs the most recent version. To see other version currently
sourced run binman::list_versions("seleniumserver")}

\item{geckover}{what version of Gecko driver to run. Default = "latest"
which runs the most recent version. To see other version currently
sourced run binman::list_versions("geckodriver"), A value of NULL
excludes adding the firefox browser to Selenium Server.}
}
\value{
list of genomic sequences referent to the urls/coordinates given by
the user and can be used to write .fasta files using the function \code{write_fasta}.
}
\description{
This function use RSelenium to open NCBI urls and extract the
fasta sequence from them. WARNING! It will automatically open the browser
chosen to retrieve the information.
}
\examples{
\dontrun{
weblist <- list(
  target1 =
    "https://www.ncbi.nlm.nih.gov/nuccore/NC_010443.5?report=fasta&from=39157155&to=39157675",
  target2 =
    "https://www.ncbi.nlm.nih.gov/nuccore/NC_010458.4?report=fasta&from=27277683&to=27278206",
  target3 =
    "https://www.ncbi.nlm.nih.gov/nuccore/NC_010445.4?report=fasta&from=46206177&to=46206699"
)
#' get_fasta(weblist = weblist, browser = "chrome", port = 4455L )
}
}
\seealso{
\itemize{
\item \code{\link[=write_fasta]{write_fasta()}} write .fasta files from a character list of sequences.
\item \code{\link[=get_coord_website]{get_coord_website()}} use genomic coordinates to find the associated NCBI urls.
\item \code{\link[=tag_fasta]{tag_fasta()}} mark the target region for the off target region.
For rhAmpSeq, panel design.
}
}
